some Role refactoring
