+++ src/pages/Pools/Home/Roles/types.ts
@@ -1,0 +1,9 @@
+// Copyright 2022 @paritytech/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export type RoleEdit = {
+  oldAddress: string;
+  newAddress: string;
+  valid: boolean;
+  reformatted: boolean;
+};
